requirements.txt
Django==5.2
djangorestframework==3.14.0
django-cors-headers==4.3.1
numpy==1.26.4
pandas==2.2.1
scikit-learn==1.4.1
spacy==3.7.4
networkx==3.2.1
python-dotenv==1.0.1
selenium==4.18.1
webdriver-manager==4.0.1 